Move :protoarray to :ethereum:forkchoice
Renaming from just protoarray to forkchoice because protoarray is the way we implement fork choice from the spec.
