+++ data/provider/build.gradle
@@ -4,6 +4,7 @@ dependencies {
     implementation project(':data')
     implementation project(':data:serializer')
     implementation project(':ethereum:core')
+    implementation project(':ethereum:forkchoice')
     implementation project(':ethereum:spec')
     implementation project(':ethereum:statetransition')
     implementation project(':infrastructure:async')
@@ -12,7 +13,6 @@ dependencies {
     implementation project(':networking:eth2')
     implementation project(':infrastructure:serviceutils')
     implementation project(':infrastructure:ssz')
-    implementation project(':protoarray')
     implementation project(':storage')
     implementation project(':beacon:sync')
     implementation project(':validator:api')
