+++ fork-choice-tests/build.gradle
@@ -3,10 +3,10 @@ jar { enabled = false }
 dependencies {
     integrationTestImplementation project(':data')
     integrationTestImplementation project(':ethereum:core')
+    integrationTestImplementation project(':ethereum:forkchoice')
     integrationTestImplementation testFixtures(project(':ethereum:spec'))
     integrationTestImplementation project(':ethereum:statetransition')
     integrationTestImplementation project(':eth-tests')
-    integrationTestImplementation project(':protoarray')
     integrationTestImplementation project(':infrastructure:bls')
     integrationTestImplementation project(':infrastructure:ssz')
     integrationTestImplementation project(':storage')
