--- protoarray/src/main/java/tech/pegasys/teku/protoarray/package-info.java
@@ -1,10 +1,0 @@
-/**
- * This package provides an implementation of the ProtoArray algorithm, which is an optimization to
- * the naive Fork Choice algorithm described in Ethereum 2.0 specifications, authored by Diederik
- * Loerakker.
- *
- * <p>The code is mostly translated from the Apache 2.0 licensed Lighthouse codebase. Thanks to Paul
- * Hauner and others for the implementation.
- * https://github.com/sigp/lighthouse/tree/master/eth2/proto_array_fork_choice
- */
-package tech.pegasys.teku.protoarray;
