+++ services/beaconchain/build.gradle
@@ -6,6 +6,7 @@ dependencies {
   implementation project(':data:provider')
   implementation project(':ethereum:core')
   implementation project(':ethereum:events')
+  implementation project(':ethereum:forkchoice')
   implementation project(':ethereum:networks')
   implementation project(':ethereum:spec')
   implementation project(':ethereum:statetransition')
@@ -19,7 +20,6 @@ dependencies {
   implementation project(':networking:p2p')
   implementation project(':networking:eth2')
   implementation project(':pow')
-  implementation project(':protoarray')
   implementation project(':infrastructure:serviceutils')
   implementation project(':services:powchain')
   implementation project(':services:timer')
