+++ settings.gradle
@@ -7,6 +7,7 @@ include 'ethereum:spec'
 include 'ethereum:dataproviders'
 include 'ethereum:events'
 include 'ethereum:executionlayer'
+include 'ethereum:forkchoice'
 include 'ethereum:networks'
 include 'ethereum:pow:api'
 include 'ethereum:statetransition'
@@ -47,7 +48,6 @@ include 'services:chainstorage'
 include 'services:timer'
 include 'teku'
 include 'pow'
-include 'protoarray'
 include 'infrastructure:ssz'
 include 'infrastructure:ssz:generator'
 include 'storage'
