+++ storage/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
   implementation project(':ethereum:core')
   implementation project(':ethereum:dataproviders')
+  implementation project(':ethereum:forkchoice')
   implementation project(':ethereum:pow:api')
   implementation project(':ethereum:spec')
   implementation project(':infrastructure:async')
@@ -13,7 +14,6 @@ dependencies {
   implementation project(':infrastructure:metrics')
   implementation project(':infrastructure:ssz')
   implementation project(':pow')
-  implementation project(':protoarray')
   implementation project(':storage:api')
 
   implementation 'com.fasterxml.jackson.core:jackson-databind'
@@ -33,10 +33,10 @@ dependencies {
   testImplementation testFixtures(project(':infrastructure:bls'))
   testImplementation testFixtures(project(':infrastructure:metrics'))
   testImplementation testFixtures(project(':ethereum:core'))
+  testImplementation testFixtures(project(':ethereum:forkchoice'))
   testImplementation project(':ethereum:networks')
   testImplementation testFixtures(project(':ethereum:spec'))
   testImplementation testFixtures(project(':infrastructure:async'))
-  testImplementation testFixtures(project(':protoarray'))
   testImplementation testFixtures(project(':pow'))
   testImplementation testFixtures(project(':storage'))
 
@@ -50,12 +50,12 @@ dependencies {
   testFixturesImplementation testFixtures(project(':infrastructure:bls'))
   testFixturesImplementation testFixtures(project(':infrastructure:metrics'))
   testFixturesImplementation testFixtures(project(':ethereum:core'))
+  testFixturesImplementation testFixtures(project(':ethereum:forkchoice'))
   testFixturesImplementation testFixtures(project(':ethereum:spec'))
   testFixturesImplementation project(':ethereum:networks')
   testFixturesImplementation testFixtures(project(':infrastructure:async'))
   testFixturesImplementation project(':infrastructure:events')
   testFixturesImplementation testFixtures(project(':pow'))
-  testFixturesImplementation testFixtures(project(':protoarray'))
   testFixturesImplementation 'org.mockito:mockito-core'
   testFixturesImplementation 'org.junit.jupiter:junit-jupiter-api'
   testFixturesImplementation 'org.junit.jupiter:junit-jupiter-params'
