+++ teku/build.gradle
@@ -12,6 +12,7 @@ dependencies {
   implementation project(':data:yaml')
   implementation project(':ethereum:core')
   implementation project(':ethereum:dataproviders')
+  implementation project(':ethereum:forkchoice')
   implementation project(':ethereum:networks')
   implementation project(':ethereum:pow:api')
   implementation project(':ethereum:spec')
@@ -27,7 +28,6 @@ dependencies {
   implementation project(':networking:p2p')
   implementation project(':networking:eth2')
   implementation project(':networking:nat')
-  implementation project(':protoarray')
   implementation project(':pow')
   implementation project(':services:beaconchain')
   implementation project(':services:chainstorage')
